+++ docs/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project does not adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 This project uses MCVERSION-MAJORMOD.MAJORAPI.MINOR.PATCH.
 
+## [1.18.2-4.3.0.0] - 2022.04.13
+### Changed
+- Updated to and requires Forge 40.0.47+
+- Mods can now retrieve the proper nutrition and saturation modifier amounts from sandwiches through `IForgeItem#getFoodProperties(ItemStack, LivingEntity)`
+
 ## [1.18.2-4.2.0.2] - 2022.03.03
 ### Changed
 - Updated to Minecraft 1.18.2